+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/advancedsettings/AndroidAdvancedSettings.kt
@@ -7,7 +7,7 @@ import androidx.activity.viewModels
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import co.electriccoin.zcash.ui.MainActivity
-import co.electriccoin.zcash.ui.common.model.WalletRestoringState
+import co.electriccoin.zcash.ui.common.model.TopAppBarSubTitleState
 import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
 import co.electriccoin.zcash.ui.screen.advancedsettings.view.AdvancedSettings
 
@@ -21,7 +21,7 @@ internal fun MainActivity.WrapAdvancedSettings(
 ) {
     val walletViewModel by viewModels<WalletViewModel>()
 
-    val walletRestoringState = walletViewModel.walletRestoringState.collectAsStateWithLifecycle().value
+    val walletState = walletViewModel.walletStateInformation.collectAsStateWithLifecycle().value
 
     WrapAdvancedSettings(
         goBack = goBack,
@@ -29,7 +29,7 @@ internal fun MainActivity.WrapAdvancedSettings(
         goExportPrivateData = goExportPrivateData,
         goChooseServer = goChooseServer,
         goSeedRecovery = goSeedRecovery,
-        walletRestoringState = walletRestoringState,
+        topAppBarSubTitleState = walletState
     )
 }
 
@@ -41,7 +41,7 @@ private fun WrapAdvancedSettings(
     goChooseServer: () -> Unit,
     goSeedRecovery: () -> Unit,
     goDeleteWallet: () -> Unit,
-    walletRestoringState: WalletRestoringState,
+    topAppBarSubTitleState: TopAppBarSubTitleState,
 ) {
     BackHandler {
         goBack()
@@ -53,6 +53,6 @@ private fun WrapAdvancedSettings(
         onExportPrivateData = goExportPrivateData,
         onChooseServer = goChooseServer,
         onSeedRecovery = goSeedRecovery,
-        walletRestoringState = walletRestoringState,
+        topAppBarSubTitleState = topAppBarSubTitleState,
     )
 }