+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/sendconfirmation/AndroidSendConfirmation.kt
@@ -26,7 +26,7 @@ import cash.z.ecc.android.sdk.model.ZecSend
 import co.electriccoin.zcash.spackle.Twig
 import co.electriccoin.zcash.ui.MainActivity
 import co.electriccoin.zcash.ui.R
-import co.electriccoin.zcash.ui.common.model.WalletRestoringState
+import co.electriccoin.zcash.ui.common.model.TopAppBarSubTitleState
 import co.electriccoin.zcash.ui.common.viewmodel.AuthenticationViewModel
 import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
 import co.electriccoin.zcash.ui.design.component.CircularScreenProgressIndicator
@@ -72,7 +72,7 @@ internal fun MainActivity.WrapSendConfirmation(
 
     val supportMessage = supportViewModel.supportInfo.collectAsStateWithLifecycle().value
 
-    val walletRestoringState = walletViewModel.walletRestoringState.collectAsStateWithLifecycle().value
+    val walletState = walletViewModel.walletStateInformation.collectAsStateWithLifecycle().value
 
     WrapSendConfirmation(
         activity = this,
@@ -86,7 +86,7 @@ internal fun MainActivity.WrapSendConfirmation(
         spendingKey = spendingKey,
         supportMessage = supportMessage,
         synchronizer = synchronizer,
-        walletRestoringState = walletRestoringState,
+        topAppBarSubTitleState = walletState,
     )
 }
 
@@ -105,7 +105,7 @@ internal fun WrapSendConfirmation(
     spendingKey: UnifiedSpendingKey?,
     supportMessage: SupportInfo?,
     synchronizer: Synchronizer?,
-    walletRestoringState: WalletRestoringState
+    topAppBarSubTitleState: TopAppBarSubTitleState,
 ) {
     val scope = rememberCoroutineScope()
 
@@ -205,7 +205,7 @@ internal fun WrapSendConfirmation(
                         }
                 }
             },
-            walletRestoringState = walletRestoringState
+            topAppBarSubTitleState = topAppBarSubTitleState,
         )
 
         if (sendFundsAuthentication.value) {